✨ [feat] #104 가이드 이후 촬영 라이브 인식을 사람+사물로 확장
