+++ Chalkak/Core/Guide/Distance/BoundingBoxManager.swift
@@ -11,30 +11,53 @@ import Vision
 class BoundingBoxManager: NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
     var onMultiBoundingBoxUpdate: (([CGRect]) -> Void)?
 
+    private var frameCounter = 0
+    /// 매 N프레임마다 한 번씩만 인식 (30fps 입력 → ≈10fps 인식)
+    /// 촬영/프리뷰 프레임에는 영향 없음
+    private let frameInterval = 3
+
     func captureOutput(_ output: AVCaptureOutput,
                        didOutput sampleBuffer: CMSampleBuffer,
                        from connection: AVCaptureConnection) {
+        frameCounter &+= 1
+        guard frameCounter % frameInterval == 0 else { return }
+
         guard let pixelBuffer = CMSampleBufferGetImageBuffer(sampleBuffer) else { return }
 
-        let request = VNDetectHumanRectanglesRequest { [weak self] req, _ in
-            guard let results = req.results as? [VNHumanObservation] else {
-                DispatchQueue.main.async {
+        let request = VNGenerateForegroundInstanceMaskRequest()
+        let handler = VNImageRequestHandler(cvPixelBuffer: pixelBuffer,
+                                            orientation: .right,
+                                            options: [:])
+        do {
+            try handler.perform([request])
+
+            guard let result = request.results?.first as? VNInstanceMaskObservation else {
+                DispatchQueue.main.async { [weak self] in
                     self?.onMultiBoundingBoxUpdate?([])
                 }
                 return
             }
 
-            let boxes = results.map { $0.boundingBox }
-            DispatchQueue.main.async {
+            let maskedBuffer = try result.generateMaskedImage(
+                ofInstances: result.allInstances,
+                from: handler,
+                croppedToInstancesExtent: false
+            )
+
+            let boxes: [CGRect]
+            if let unionBox = unionBoundingBox(fromMaskedBuffer: maskedBuffer) {
+                boxes = [unionBox]
+            } else {
+                boxes = []
+            }
+
+            DispatchQueue.main.async { [weak self] in
                 self?.onMultiBoundingBoxUpdate?(boxes)
             }
+        } catch {
+            DispatchQueue.main.async { [weak self] in
+                self?.onMultiBoundingBoxUpdate?([])
+            }
         }
-
-        request.upperBodyOnly = true
-
-        let handler = VNImageRequestHandler(cvPixelBuffer: pixelBuffer,
-                                            orientation: .right,
-                                            options: [:])
-        try? handler.perform([request])
     }
 }
