+++ Chalkak/Core/Guide/Overlay/OverlayManager.swift
@@ -74,7 +74,7 @@ class OverlayManager: ObservableObject {
 
                 /// 4단계: 마스크 알파 채널에서 union BoundingBox 추출
                 let unionBoxes: [CGRect]
-                if let unionBox = self.unionBoundingBox(from: maskedPixelBuffer) {
+                if let unionBox = unionBoundingBox(fromMaskedBuffer: maskedPixelBuffer) {
                     let correctedBox: CGRect
                     if isFront {
                         // 좌우 반전
@@ -116,47 +116,6 @@ class OverlayManager: ObservableObject {
                 DispatchQueue.main.async { completion() }
             }
         }
-    }
-
-    //MARK: - 마스크 union bounding box 계산
-    /// 마스크된 BGRA 이미지의 알파 채널에서 non-zero 픽셀 합집합 영역을 Vision 정규화 좌표(원점 좌하단)로 반환
-    private func unionBoundingBox(from maskedBuffer: CVPixelBuffer) -> CGRect? {
-        let width = CVPixelBufferGetWidth(maskedBuffer)
-        let height = CVPixelBufferGetHeight(maskedBuffer)
-        let bytesPerRow = CVPixelBufferGetBytesPerRow(maskedBuffer)
-
-        CVPixelBufferLockBaseAddress(maskedBuffer, .readOnly)
-        defer { CVPixelBufferUnlockBaseAddress(maskedBuffer, .readOnly) }
-
-        guard let base = CVPixelBufferGetBaseAddress(maskedBuffer) else { return nil }
-        let bytes = base.assumingMemoryBound(to: UInt8.self)
-
-        var minX = width
-        var minY = height
-        var maxX = -1
-        var maxY = -1
-
-        // BGRA 픽셀의 알파 채널(offset +3)이 0보다 크면 전경
-        for y in 0..<height {
-            let rowStart = y * bytesPerRow
-            for x in 0..<width where bytes[rowStart + x * 4 + 3] > 0 {
-                if x < minX { minX = x }
-                if x > maxX { maxX = x }
-                if y < minY { minY = y }
-                if y > maxY { maxY = y }
-            }
-        }
-
-        guard maxX >= 0, maxY >= 0 else { return nil }
-
-        let w = CGFloat(width)
-        let h = CGFloat(height)
-        return CGRect(
-            x: CGFloat(minX) / w,
-            y: 1.0 - CGFloat(maxY + 1) / h,
-            width: CGFloat(maxX - minX + 1) / w,
-            height: CGFloat(maxY - minY + 1) / h
-        )
     }
 
     //MARK: - 실루엣 오버레이 이미지 생성
